requirements.txt
future==0.18.0
hydra-core>=1.1
loguru>=0.3
numpy>=1.17
omegaconf>=2.1
scikit-learn>=1.0
tqdm>=4.50.0
faiss-gpu==1.7.2
jupyterlab==3.5.1
pandas
matplotlib
seaborn
easydict
einops
pytorch-lightning
lightning-bolts
lightly
timm
ftfy
kornia
tensorflow-datasets
tensorboard==2.8.0
tensorflow-cpu
git+https://github.com/openai/CLIP.git
opencv-python==4.6.0.66
imagesize==1.4.1
